src/GovUk.Education.ExploreEducationStatistics.Admin/Controllers/Api/Statistics/DataController.cs
using System;
using System.Diagnostics;
using System.Threading.Tasks;
using System.Web.Http;
using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
using GovUk.Education.ExploreEducationStatistics.Data.Services.Interfaces;
using GovUk.Education.ExploreEducationStatistics.Data.Services.ViewModels;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Logging;

namespace GovUk.Education.ExploreEducationStatistics.Admin.Controllers.Api.Statistics
{
    [Route("api/data/[controller]")]
    [ApiController]
    [Authorize]
    public class DataController : ControllerBase
    {
        private readonly IDataService<ResultWithMetaViewModel> _dataService;
        private readonly ILogger _logger;

        public DataController(IDataService<ResultWithMetaViewModel> dataService,
            ILogger<DataController> logger)
        {
            _dataService = dataService;
            _logger = logger;
        }

        [HttpPost]
        public Task<ActionResult<ResultWithMetaViewModel>> Query([FromUri] Guid releaseId,
            [FromBody] ObservationQueryContext query)
        {
            var stopwatch = Stopwatch.StartNew();
            stopwatch.Start();

            var resultWithMetaViewModel = _dataService.Query(query, releaseId);

            stopwatch.Stop();
            _logger.LogDebug("Query {Query} executed in {Time} ms", query, stopwatch.Elapsed.TotalMilliseconds);

            return resultWithMetaViewModel.HandleFailuresOrOk();
        }
    }
}